requirements.txt
pytest>=8.0.0
pytest-json-report>=1.5.0
pytest-html>=2.0.0
google-cloud-firestore==2.13.0
requests>=2.31.0
